#1 Session Stats are not updated
The error happens because in `Server.incStats()` we get a copy of the
`sessionStats` slice's `sessionStats` struct.
We then modify this copy which does not affect the `sessionStats`
struct residing in slice `Server.sessionStats`.

With this solution `Server.sessionStats` slice will store references of
`sessionStats` structs.
Thus when iterating the slice we will get a reference to the struct
rather than a copy and we will modify the slice's struct.