#1 Session Stats are not updated
The error happens because in `Server.incStats()` we get a copy of the
`sessionStats` slice's `sessionStats` struct.
We then modify this copy which does not affect the `sessionStats`
struct residing in slice `Server.sessionStats`.

With this solution in `Server.incStats()` we will use index iterating
in order to affect the slice's struct instead of its copy.