#5 use default CheckOrigin
As it currently stands, `CheckOrigin` when doing a websocket upgrade
does not provide any check regarding the `Origin` header.
Solution to this is to use the default `CheckOrigin` provided by the
`websocket` module, which checks the `Origin` header against the `Host`
header.

This function could be enhanced according to requirements on a new
feature request.